backend/requirements.txt
django
djangorestframework>=3.12.0
djangorestframework-simplejwt>= 5.2.2
django-debug-toolbar==4.4.6
django-cors-headers
psycopg2-binary
virtualenv
python-dotenv
django-storages
boto3
cryptography
django-rest-passwordreset
Pillow
pydub
vosk
pandas
textblob
transformers
gunicorn